pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nou-ai"
version = "0.1.0"
description = "Reverse the attack. 4-stage LLM guardrail pipeline for prompt injection defense."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
